Add logging and validation to data generation script
It is now completely unreadable.

I wish we had a more elegant way of representing error states while not compromising on utility. There's a lot of
```ts
let foo;
try {
  // [...]
} catch // [...]
```
that completely breaks up the usual reading flow of `const foo = // [...]`.

I do realise the answer here is something like `Result`, but I don't think adding a version of that in here is going to help, especially as it's less common in Javascript land. Maybe a future project could use it.

Mini-rant over. Why are you even reading this? Why am _I_ expecting someone to read this?