Improve gradient generation to generate extra stops when changing hues
This was a fun little thing to figure out. This commit tries to smoothen out the background gradients by adding in extra stops. The colours for each intermediate stop are generated by converting the colours to OKLab (technically OKLch), which has a better perceptual model of colour (e.g. a straight line through the colour space should appear to change uniformly and not change with brightness/saturation unexpectedly). It then makes a number of stops, based on the hue difference, and linearly interpolates some extra points.

This may need some tweaking (or tearing out) in the future. Generating smoother curves, potentially, though this is good enough without being (too) computationally expensive. The number of initial colours should be low to start with.