Remove header
It was actually quite hard to make this decision. Having branding in a top harder is extremely ubiquitous, to the point I was questioning whether people would get lost without it. 
I quickly realised that was a weird idea. People today know how to use the internet. They'll be fine without that screen space being used. Besides, it gives an opportunity to focus on what's truly important here: the content.