Add analytics
Hello person from the future reding this.
I self-host an instance of [Umami](https://umami.is/), which is an open-source analytics thing that does not use cookies. I have also configured it to respect the Do Not Track header, in case your browser sends that.
This instance is hosted at https://analytics.sthom.kiwi. Any good adblocker will see "analytics" in the URL and block it without a second thought.