Write summary and content for different gender identities
As an aside, I'd like to thank Kae Anthony of Majestic Mess for all the work they did on their site, which appears to be down now. I really appreciate the call-out that a lot of information doesn't appear to have a primary source, and they did a lot of work compiling information similarly to what I'm trying to do now. Their email interviews with flag creators were a treat to read. The interview with Marilyn Roxie (who I'd also like to acknowledge in this commit message) about the GQ flag was especially detailed with lots of good background. I suspect Marilyn's studies in library tech helped there.

I stand on the shoulders of giants, but these giants are also human and I want to thank them for what they did before me.